internal/processor: add batch proc for db events
As described in the architecture of the indexer, the system has a
processor component which takes events from the Updates Queue component
(in this case, the InMemoryQueue) and applies those events to the
database.
For a clean way to close the program, a context channel is used to be
able to send a Close Signal to the processor.